backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
gunicorn==21.2.0
requests==2.31.0
PyJWT==2.10.1
bcrypt==5.0.0
python-dotenv==1.0.0
schedule==1.2.0
APScheduler==3.10.1
